Remove buddy allocator from poptrie.c
This commit keeps the pre-allocation of nodes and leaves but does not
instantiate the buddy allocators for them. Instead, simple get_new_node()
and get_new_leaf() functions just return the next index of a pre-allocated
node.

The replaced code only works for adding a route -- more calls to
buddy_alloc2() need to replaced, and all buddy_free2() calls need to
be replaced. Once they're replaced, more tests from tests/basic.c can
be uncommented.

Eventually, we can replace the pre-allocation of nodes and leaves up-front
by using the slab allocator in the kernel. We may also make base0 and base1
directly reference a memory address instead of an index in a pre-allocated
array, but that would involve expanding nodes to 32 bytes instead of 24.